feat(arch): audit dependency-injection-architecture.md
- Refactor container.py (2307 lines) into modular src/core/container/ (8 modules)
- Update doc sections 0, 3.2, 4.1, 9.1, 9.2 to match actual implementation
- Add tests for get_cache, get_event_bus, get_encryption_service
- Create docs/guides/dependency-injection-usage.md practical guide
- Add Architecture Compliance Audit Process to WARP.md

Container modules:
- infrastructure.py: cache, db, secrets, logging, encryption
- events.py: event bus with 27 handler subscriptions
- repositories.py: all repository factories
- auth_handlers.py: 15 auth handlers
- provider_handlers.py: 5 provider handlers
- data_handlers.py: 9 account/transaction handlers
- providers.py: get_provider factory
- authorization.py: Casbin RBAC

F6.4 Document 2/23 complete
